fixed a few small spec issues
