+++ src/Protocol/Client/Capabilities/DocumentSymbolCapability.cs
@@ -27,6 +27,6 @@ namespace OmniSharp.Extensions.LanguageServer.Protocol.Client.Capabilities
         /// </summary>
         [Obsolete(Constants.Proposal)]
         [Optional]
-        public TagSupportCapability TagSupport { get; set; }
+        public Supports<TagSupportCapability> TagSupport { get; set; }
     }
 }
